refactor main.rs, splitting into multiple files
