+++ src/main.rs
@@ -6,13 +6,12 @@
 use anyhow::*;
 use async_std::{
     net::UdpSocket,
-    process::{Command, Stdio},
+    process::Command,
     sync::{Arc, Barrier, RwLock},
     task::{sleep, spawn},
 };
-use serde::Serialize;
 use serde_json::json;
-use std::{collections::HashMap, env, process::exit, time::Duration};
+use std::{collections::HashMap, env, process::exit};
 
 mod config;
 use config::*;
@@ -20,73 +19,14 @@ use config::*;
 mod rusage;
 use rusage::*;
 
-#[derive(Serialize, Clone)]
-#[serde(untagged)]
-enum MetricValue {
-    Str(String),
-    Num(f64),
-    Arr(Vec<HashMap<String, MetricValue>>),
-}
+mod subproc;
+use subproc::*;
 
-impl MetricValue {
-    fn as_f64(self) -> f64 {
-        match self {
-            Self::Num(x) => x,
-            _ => panic!("not an f64"),
-        }
-    }
-}
+mod statsd;
+use statsd::*;
 
-impl From<String> for MetricValue {
-    fn from(string: String) -> Self {
-        MetricValue::Str(string)
-    }
-}
-
-macro_rules! num_type {
-    ($type:ty) => {
-        impl From<$type> for MetricValue {
-            fn from(num: $type) -> Self {
-                MetricValue::Num(num as f64)
-            }
-        }
-    };
-}
-num_type!(i32);
-num_type!(i64);
-num_type!(f64);
-
-async fn statsd_listener(barrier: Arc<Barrier>, statsd_buf: Arc<RwLock<String>>) -> Result<String> {
-    let socket = UdpSocket::bind("127.0.0.1:8125").await;
-    let socket = match socket {
-        Ok(s) => s,
-        Err(error) => panic!("Cannot bind to 127.0.0.1:8125: {}", error),
-    };
-    barrier.wait().await; // indicates to main task that socket is listening
-
-    loop {
-        let mut buf = vec![0u8; 4096];
-        let (recv, _peer) = socket.recv_from(&mut buf).await?;
-
-        let datum = String::from_utf8(buf[..recv].into()).unwrap_or_else(|_| String::new());
-        statsd_buf.write().await.push_str(&datum);
-    }
-}
-
-fn get_statsd_metrics(metrics: &mut HashMap<String, MetricValue>, udp_data: String) -> Result<()> {
-    let lines = udp_data.trim().lines();
-    for line in lines {
-        let metric: Vec<&str> = match line.split('|').next() {
-            None => continue,
-            Some(metric) => metric.split(':').collect(),
-        };
-        if metric.len() < 2 {
-            continue;
-        }
-        metrics.insert(metric[0].into(), metric[1].parse::<f64>()?.into());
-    }
-    Ok(())
-}
+mod metric_value;
+use metric_value::*;
 
 fn get_kernel_metrics(wall_time: f64, data: Rusage, metrics: &mut HashMap<String, MetricValue>) {
     metrics.insert("max.res.size".into(), data.max_res_size.into());
@@ -95,61 +35,6 @@ fn get_kernel_metrics(wall_time: f64, data: Rusage, metrics: &mut HashMap<String
 
     let pct = (data.user_time + data.system_time) * 100.0 / wall_time;
     metrics.insert("cpu.pct.wall.time".into(), pct.into());
-}
-
-fn get_stdio() -> Stdio {
-    match env::var("SIRUN_NO_STDIO") {
-        Ok(_) => Stdio::null(),
-        Err(_) => Stdio::inherit(),
-    }
-}
-
-async fn run_setup_or_teardown(typ: &str, config: &Config) -> Result<()> {
-    if env::var("SIRUN_SKIP_SETUP").is_ok() {
-        return Ok(());
-    }
-    let command_arr = if typ == "setup" {
-        &config.setup
-    } else {
-        &config.teardown
-    };
-    let command_arr = match command_arr {
-        Some(command_arr) => command_arr,
-        None => return Ok(()),
-    };
-    let env = &config.env;
-    let mut code: i32 = 1;
-    let mut attempts: u8 = 0;
-    while code != 0 {
-        if attempts == 100 {
-            bail!("{} script did not complete successfully. aborting.", typ);
-        }
-        let command = command_arr[0].clone();
-        let args = command_arr.iter().skip(1);
-        code = Command::new(command)
-            .args(args)
-            .envs(env.clone())
-            .stdout(get_stdio())
-            .stderr(get_stdio())
-            .status()
-            .await?
-            .code()
-            .expect("no exit code");
-        if code != 0 {
-            sleep(Duration::from_secs(1)).await;
-            attempts += 1;
-        }
-    }
-
-    Ok(())
-}
-
-async fn run_setup(config: &Config) -> Result<()> {
-    run_setup_or_teardown("setup", config).await
-}
-
-async fn run_teardown(config: &Config) -> Result<()> {
-    run_setup_or_teardown("teardown", config).await
 }
 
 async fn test_timeout(timeout: u64) {
@@ -163,17 +48,9 @@ async fn run_test(config: &Config, mut metrics: &mut HashMap<String, MetricValue
         spawn(test_timeout(timeout));
     }
 
-    let command = config.run[0].clone();
-    let args = config.run.iter().skip(1);
     let start_time = std::time::Instant::now();
     let rusage_start = Rusage::new();
-    let status = Command::new(command)
-        .args(args)
-        .envs(&config.env)
-        .stdout(get_stdio())
-        .stderr(get_stdio())
-        .status()
-        .await?;
+    let status = run_cmd(&config.run, &config.env).await?;
     let duration = start_time.elapsed().as_micros();
     let rusage_result = Rusage::new() - rusage_start;
     metrics.insert("wall.time".to_owned(), (duration as f64).into());
@@ -192,23 +69,15 @@ async fn run_iteration(
 ) -> Result<HashMap<String, MetricValue>> {
     run_setup(&config).await?;
 
-    let mut metrics = HashMap::new();
     let mut sub_config: Config = config.clone();
     let json_config = serde_yaml::to_string(&config)?;
     sub_config.env.insert("SIRUN_ITERATION".into(), json_config);
-    let command = env::args().next().unwrap();
-    let status = Command::new(command)
-        .envs(&sub_config.env)
-        .stdout(get_stdio())
-        .stderr(get_stdio())
-        .status()
-        .await?;
+    let status = run_cmd(&env::args().take(1).collect(), &sub_config.env).await?;
     let status = status.code().expect("no exit code");
     if status != 0 && status <= 128 {
         exit(status);
     }
-    get_statsd_metrics(&mut metrics, statsd_buf.read().await.clone())?;
-    statsd_buf.write().await.clear();
+    let metrics = get_statsd_metrics(statsd_buf).await?;
 
     run_teardown(&config).await?;
